Reorganize ref and parameter tests
